{{ cookiecutter.project_repo }}/requirements.txt
# external requirements
click
Sphinx
coverage
awscli>=1.29.35
flake8
python-dotenv>=0.5.1
requests
requests-mock
scipy
geosyspy
pandas
shapely
rasterio
xarray
pyproj
matplotlib
fastapi
jsonschema
hypercorn[trio]==0.14.3
